% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sediments_Corse-package.R
\docType{package}
\name{Sediments_Corse-package}
\alias{Sediments_Corse}
\alias{Sediments_Corse-package}
\title{Sediments_Corse: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/guilhemmarre/Sediments_Corse}
  \item Report bugs at \url{https://github.com/guilhemmarre/Sediments_Corse/issues}
}

}
\author{
\strong{Maintainer}: Guilhem Marre \email{guilhem.marre@andromede-ocean.com} (\href{https://orcid.org/0000-0001-5280-7015}{ORCID}) [copyright holder]

}
\keyword{internal}
